__all__ =======
Show how to use __all__ in modules so that in __init__.py we
can simply use from .module import *

Add utils with @export decorator for easy management of
__all__ in modules.

Delete modules in __init__.py so that they do not show up
in the packages namespace. This is somewhat ugly but I
do not know how to solve this problem in a cleaner fashion.
One would expect that something that is not presentt in the
namespace of __init__.py will not show up in the namespace of
the package (but this is what happens).